5. Stuck - Need Property Changed Events
Going back to having some kind of singleton encapsulating the company
file and exterior info.

Need to figure out how to do property changed events properly so that I
can save the data to the file when any change is made. Changes should
cascade up if they're an Employee change.