9. Huge UI Improvement
Overhauled the UI again. Got rid of the SetupWindow, and now it goes
directly to the MainWindow, which has a Menu for opening, saving,
closing, etc. Looks much better .

Also spent a bunch of time messing with buttons and images, bitmaps, and
other shit. Best way to do it now is just to use an image, and use
OnLeftMousebuttonUp, or something like that.

Also explored a lot of customization options as far as appearance within
WPF. Very cool.