10. Code Restructureing
Changed the AppManager from a singleton that contains the company and
fileStream, to being a helper class that does some loading/saving.

Changed a lot of code in MainWindow.xaml.cs.

TODO: I left off on trying to make the clock in/out buttons go from gray
to black and vice versa when the employee selection is changed.